requirements.txt
asgiref==3.7.2
Django==4.2.2
django-environ==0.10.0
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
gunicorn==20.1.0
djoser==2.2.0
psycopg2-binary==2.9.6
drf-spectacular==0.26.5
django-cors-headers==4.2.0
drf-extra-fields==3.7.0
django-redis==5.2.0
markdown==3.0.0
django-filter==23.3
django-phonenumber-field==7.2.0
phonenumberslite==8.13.23
django-extensions==3.2.3
django-smsru==1.0.4
